+++ ModelSwapperSkins/ModelInfos/ModelInfoInitializer.cs
@@ -124,6 +124,7 @@ namespace ModelSwapperSkins.ModelInfos
             SetModelInfo("ClayGrenadierBody", new ModelInfo(3f));
             SetModelInfo("CrocoBody", new ModelInfo(22.5f));
             SetModelInfo("FlyingVerminBody", new ModelInfo(5f));
+            SetModelInfo("GolemBody", new ModelInfo(5.5f));
             SetModelInfo("GrandParentBody", new ModelInfo(50f));
             SetModelInfo("HereticBody", new ModelInfo(4f));
             SetModelInfo("ImpBody", new ModelInfo(2f));