+++ CHANGELOG.md
@@ -13,6 +13,7 @@
   * Clay Templar model scale and head position
   * Acrid model scale
   * Blind Pest feet rotation
+  * Stone Golem model scale and particles
 
 **1.0.0 Changes:**